moving mapping logic to type_directory.go
